check status on $m->abort and handle it
Result of $m->abort('xxx') is returned to caller, so
it can be our status.

When we get status or just return from request headers can be
sent allready. In this case it's too late to send custom status
line and we warn about it, unless status is 200.

If headers are not sent then it means we have response without
any content. In this case we send status if it's defined or
"204: No content".